Refactor calender components
