+++ packages/components/src/CalendarDateRange/types.ts
@@ -1,6 +1,6 @@
-import type { CalendarDateProps } from '../types';
+import type { CalendarBaseProps, CalendarDateProps } from '../types';
 
-export interface CalendarDateRangeProps extends CalendarDateProps {
+export interface CalendarDateRangeProps extends CalendarDateProps, CalendarBaseProps {
 	startDate: Date | string;
 	endDate: Date | string;
 }
